requirements.txt
asgiref
attrs
blinker

certifi
chardet
click
colorama
crispy-bootstrap5
cryptography
defer
distlib
distro
distro-info
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
httplib2
idna

jeepney
jsonschema
jsonschema-specifications
keyring

lazr.restfulclient
lazr.uri
louis
macaroonbakery
more-itertools
netifaces

olefile
pexpect
Pillow
platformdirs
protobuf
psycopg2

pycairo
PyJWT
pymacaroons
PyNaCl
pyparsing
pyRFC3339
python-dateutil
python-debian
pytz
pyxdg
PyYAML
referencing
reportlab
requests
rpds-py
SecretStorage
six
sqlparse
virtualenv
xdg

zipp